requirements.txt
accelerate==1.13.0
annotated-doc==0.0.4
anyio==4.13.0
certifi==2026.2.25
click==8.3.2
colorama==0.4.6
einops==0.8.2
filelock==3.25.2
fsspec==2026.2.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
# huggingface_hub==1.9.0
huggingface_hub>=0.23.2,<1.0.0  # Downgrading from 1.9.0 to stable 0.x
idna==3.11
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.3
packaging==26.0
pillow==12.1.1
psutil==7.2.2
Pygments==2.20.0
PyYAML==6.0.3
regex==2026.4.4
rich==14.3.3
safetensors==0.7.0
scipy==1.17.1
setuptools==70.2.0
shellingham==1.5.4
sympy==1.14.0
timm==1.0.26
tokenizers==0.22.2
torch==2.11.0+cpu
torchaudio==2.11.0+cpu
torchvision==0.26.0+cpu
tqdm==4.67.3
# transformers==5.5.0
transformers>=4.44.2,<5.0.0  # Downgrading from 5.5.0 to stable 4.x
typer==0.24.1
typing_extensions==4.15.0
